#1497 move setValue/getValue tests from SelenideMethodsTest to a separate class GetSetValueTest
SelenideMethodsTest was too long, need to split it even more.
